launch_url [nfc]: Move launchUrl logic to a new file
- Move the existing launchUrl logic from content.dart to a new file
  named launch_url.dart.
- Split function into pieces to handle realm-based and non-realm
  URLs.
- Refactor error handling into a private  function.
